Moved email task to notifications
